+++ packages/launchpad/src/migration/RenameSupport.vue
@@ -18,13 +18,13 @@
   <BeforeAfter>
     <template #before>
       <HighlightedFilesList
-        :files="props.gql.supportFiles?.map(x => x.before) || []"
+        :files="props.gql.supportFiles?.before ? [props.gql.supportFiles.before] : []"
         highlight-class="text-red-500"
       />
     </template>
     <template #after>
       <HighlightedFilesList
-        :files="props.gql.supportFiles?.map(x => x.after) || []"
+        :files="props.gql.supportFiles?.after ? [props.gql.supportFiles.after] : []"
         highlight-class="text-jade-500"
       />
     </template>
