+++ packages/launchpad/src/migration/fragments/HighlightedFilesList.cy.tsx
@@ -1,16 +1,17 @@
 import HighlightedFilesList from './HighlightedFilesList.vue'
-import { FilePart, formatMigrationFile, regexps } from '@packages/data-context/src/util/migrationFormat'
+import { FilePart, formatMigrationFile } from '@packages/data-context/src/util/migrationFormat'
+import { regexps } from '@packages/data-context/src/sources/migration/regexps'
 
 describe('<HighlightedFilesList/>', { viewportWidth: 1119 }, () => {
   it('renders expected content', () => {
     type PropType = Readonly<Array<{ parts: readonly FilePart[] }>>
     const _files = [
-      'cypress/e2e/app.cy.js',
-      'cypress/e2e/blog-post.cy.js',
-      'cypress/e2e/homeSpec.cy.js',
-      'cypress/e2e/company.cy.js',
-      'cypress/e2e/sign-up.cy.js',
-    ].map((x) => formatMigrationFile(x, new RegExp(regexps.e2e.usingDefaultIntegrationFolder.afterRegexp)))
+      'cypress/integration/app.spec.js',
+      'cypress/integration/blog-post.spec.js',
+      'cypress/integration/homeSpec.spec.js',
+      'cypress/integration/company.spec.js',
+      'cypress/integration/sign-up.spec.js',
+    ].map((x) => formatMigrationFile(x, new RegExp(regexps.e2e.before.defaultFolderDefaultTestFiles)))
 
     const files: PropType = _files.map((x) => ({ parts: x }))
